requirements.txt
pandas
numpy
ipykernel
tweepy
requests
boto3
langchain_aws
langchain
emoji
python-dotenv

#API構築のためのライブラリ
fastapi
uvicorn

pytest